+++ actix-utils/CHANGES.md
@@ -4,6 +4,10 @@
 
 * Fix framed transport error handling
 
+* Added Clone impl for Either service
+
+* Added Service and NewService for Stream dispatcher
+
 
 ## [0.1.0] - 2018-12-09